Move reverseVehicleOrderTable to OrderManager
